Create InMemorySwordsRepository class to manage data.
Move sword_list into it's own module.
Rename modules.
