+++ data.py
@@ -1,0 +1,32 @@
+sword_list = [
+    {
+    "id": "P4nKuuuurtk4-z-nwidii-k000x",
+    "name": "Zardzewiały miecz", 
+    "damage": 10, 
+    "required_strength": 5,
+    },
+    {
+    "id": "Jaa444-ni3-chc3-tyl3-z4r4bi4c",
+    "name": "Miecz Świstaka", 
+    "damage": 20, 
+    "required_strength": 15,
+    },
+    {
+    "id": "pies3k-szcz3sl1wy",
+    "name": "Miecz strachu", 
+    "damage": 42, 
+    "required_strength": 18,
+    },
+    {
+    "id": "eeeeeeeeeeeeeeeeeeee",
+    "name": "Miecz śmierci", 
+    "damage": 48, 
+    "required_strength": 21,
+    },
+    {
+    "id": "j3st-d0w0d-n4-t0-ajdi",
+    "name": "Miecz Blizny", 
+    "damage": 85, 
+    "required_strength": 70,
+    },
+]
